chore: remove obsolete kviz prototype and merge quiz admin
