--- src/kviz/app/language-select.tsx
@@ -1,141 +1,0 @@
-import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet, SafeAreaView, ScrollView } from 'react-native';
-import { router } from 'expo-router';
-import { LinearGradient } from 'expo-linear-gradient';
-import { ArrowLeft } from 'lucide-react-native';
-import { useGame } from '@/store/game-store';
-import { Language, LANGUAGES } from '@/constants/languages';
-
-export default function LanguageSelect() {
-  const { userRole, setSelectedLanguage } = useGame();
-
-  const handleLanguageSelect = (language: Language) => {
-    setSelectedLanguage(language);
-    if (userRole === 'moderator') {
-      router.push('/moderator');
-    } else {
-      router.push('/player');
-    }
-  };
-
-  return (
-    <SafeAreaView style={styles.container}>
-      <LinearGradient
-        colors={['#667eea', '#764ba2']}
-        style={styles.gradient}
-      >
-        <View style={styles.header}>
-          <TouchableOpacity
-            style={styles.backButton}
-            onPress={() => router.back()}
-          >
-            <ArrowLeft size={24} color="white" />
-          </TouchableOpacity>
-          <Text style={styles.title}>Select Language</Text>
-          <View style={styles.placeholder} />
-        </View>
-
-        <ScrollView style={styles.scrollView} showsVerticalScrollIndicator={false}>
-          <View style={styles.content}>
-            <Text style={styles.subtitle}>
-              Choose your preferred quiz language
-            </Text>
-            
-            <View style={styles.languageGrid}>
-              {LANGUAGES.map((language) => (
-                <TouchableOpacity
-                  key={language.code}
-                  style={styles.languageCard}
-                  onPress={() => handleLanguageSelect(language)}
-                  activeOpacity={0.8}
-                >
-                  <Text style={styles.flag}>{language.flag}</Text>
-                  <Text style={styles.languageName}>{language.name}</Text>
-                  <Text style={styles.nativeName}>{language.nativeName}</Text>
-                </TouchableOpacity>
-              ))}
-            </View>
-          </View>
-        </ScrollView>
-      </LinearGradient>
-    </SafeAreaView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  gradient: {
-    flex: 1,
-  },
-  header: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: 20,
-    paddingTop: 20,
-    paddingBottom: 10,
-  },
-  backButton: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    backgroundColor: 'rgba(255, 255, 255, 0.2)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: 'white',
-  },
-  placeholder: {
-    width: 40,
-  },
-  scrollView: {
-    flex: 1,
-  },
-  content: {
-    padding: 20,
-    paddingTop: 10,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: 'rgba(255, 255, 255, 0.8)',
-    textAlign: 'center',
-    marginBottom: 30,
-  },
-  languageGrid: {
-    gap: 15,
-  },
-  languageCard: {
-    backgroundColor: 'white',
-    borderRadius: 16,
-    padding: 20,
-    flexDirection: 'row',
-    alignItems: 'center',
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 4,
-  },
-  flag: {
-    fontSize: 32,
-    marginRight: 16,
-  },
-  languageName: {
-    fontSize: 18,
-    fontWeight: '600',
-    color: '#333',
-    flex: 1,
-  },
-  nativeName: {
-    fontSize: 14,
-    color: '#666',
-  },
-});
