--- src/kviz/constants/languages.ts
@@ -1,75 +1,0 @@
-export interface Language {
-  code: string;
-  name: string;
-  nativeName: string;
-  flag: string;
-}
-
-export const LANGUAGES: Language[] = [
-  {
-    code: 'en',
-    name: 'English',
-    nativeName: 'English',
-    flag: '🇺🇸',
-  },
-  {
-    code: 'es',
-    name: 'Spanish',
-    nativeName: 'Español',
-    flag: '🇪🇸',
-  },
-  {
-    code: 'de',
-    name: 'German',
-    nativeName: 'Deutsch',
-    flag: '🇩🇪',
-  },
-  {
-    code: 'fr',
-    name: 'French',
-    nativeName: 'Français',
-    flag: '🇫🇷',
-  },
-  {
-    code: 'it',
-    name: 'Italian',
-    nativeName: 'Italiano',
-    flag: '🇮🇹',
-  },
-  {
-    code: 'pt-br',
-    name: 'Portuguese (Brazil)',
-    nativeName: 'Português (Brasil)',
-    flag: '🇧🇷',
-  },
-  {
-    code: 'zh',
-    name: 'Chinese (Mandarin)',
-    nativeName: '中文',
-    flag: '🇨🇳',
-  },
-  {
-    code: 'ja',
-    name: 'Japanese',
-    nativeName: '日本語',
-    flag: '🇯🇵',
-  },
-  {
-    code: 'ko',
-    name: 'Korean',
-    nativeName: '한국어',
-    flag: '🇰🇷',
-  },
-  {
-    code: 'hi',
-    name: 'Hindi',
-    nativeName: 'हिन्दी',
-    flag: '🇮🇳',
-  },
-  {
-    code: 'sk',
-    name: 'Slovak',
-    nativeName: 'Slovenčina',
-    flag: '🇸🇰',
-  },
-];
